+++ client/src/components/resources_page/FilterTypesSection.vue
@@ -14,7 +14,7 @@
 import { defineComponent } from 'vue';
 import ResourcesPageSection from './ResourcesPageSection.vue';
 import { ResourcesPageContent as content } from '@/assets/messages/resources';
-import { ResourcesPageStyles as styles } from '@/assets/styles/resources';
+import { ResourcesPageStyles as styles } from '../../assets/styles/style_props';
 import ImageCard from '../ImageCard.vue';
 
 /**
